Updated error handling logic, updated changelog, incremented dot-release to 1.0.6.
Some hosts will not reply with errors or not-found. Sometimes, the reply
will be an SOA record, instead of an error record for _dmarc.$domain.
When this is not the case, it is not uncommon to find other forms of TXT
records and CNAME records hosted on _dmarc.$domain ; error handling has
been adjusted to relect this in the error messaging as well as update
the diagnostic info when using the -DisplayErrors parameter.

Removed extraneous -showRecord parameter which was a vestige from
initial development and testing when Get-DMARCRecord employed a
different output scheme.